+++ pnpm-workspace.yaml
@@ -27,11 +27,11 @@ packages:
   - xwiki-platform-core/xwiki-platform-node/src/main/node/core/hierarchy/hierarchy-api
   - xwiki-platform-core/xwiki-platform-node/src/main/node/core/hierarchy/hierarchy-xwiki
   - xwiki-platform-core/xwiki-platform-node/src/main/node/core/icons
-  - xwiki-platform-core/xwiki-platform-node/src/main/node/core/link-modal/link-modal-api
-  - xwiki-platform-core/xwiki-platform-node/src/main/node/core/link-modal/link-modal-default
   - xwiki-platform-core/xwiki-platform-node/src/main/node/core/link-modal/link-modal-ui
   - xwiki-platform-core/xwiki-platform-node/src/main/node/core/link-suggest/link-suggest-api
   - xwiki-platform-core/xwiki-platform-node/src/main/node/core/link-suggest/link-suggest-xwiki
+  - xwiki-platform-core/xwiki-platform-node/src/main/node/core/link-type/link-type-api
+  - xwiki-platform-core/xwiki-platform-node/src/main/node/core/link-type/link-type-default
   - xwiki-platform-core/xwiki-platform-node/src/main/node/core/livedata/livedata-api
   - xwiki-platform-core/xwiki-platform-node/src/main/node/core/livedata/livedata-componentstore
   - xwiki-platform-core/xwiki-platform-node/src/main/node/core/livedata/livedata-ui
