+++ xwiki-platform-core/xwiki-platform-blocknote/xwiki-platform-blocknote-webjar/src/main/node/package.json
@@ -21,9 +21,9 @@
     "@xwiki/platform-dsapi": "workspace:*",
     "@xwiki/platform-editors-blocknote-headless": "workspace:*",
     "@xwiki/platform-editors-blocknote-react": "workspace:*",
-    "@xwiki/platform-link-modal-default": "workspace:*",
     "@xwiki/platform-link-suggest-api": "workspace:*",
     "@xwiki/platform-link-suggest-xwiki": "workspace:*",
+    "@xwiki/platform-link-type-default": "workspace:*",
     "@xwiki/platform-markdown-syntax-config": "workspace:*",
     "@xwiki/platform-minimal-syntax-config": "workspace:*",
     "@xwiki/platform-model-api": "workspace:*",
