+++ xwiki-platform-core/xwiki-platform-node/src/main/node/core/link-modal/link-modal-ui/package.json
@@ -42,8 +42,8 @@
     "@xwiki/platform-attachments-api": "workspace:*",
     "@xwiki/platform-document-api": "workspace:*",
     "@xwiki/platform-dsapi": "workspace:*",
-    "@xwiki/platform-link-modal-api": "workspace:*",
     "@xwiki/platform-link-suggest-api": "workspace:*",
+    "@xwiki/platform-link-type-api": "workspace:*",
     "@xwiki/platform-model-api": "workspace:*",
     "@xwiki/platform-model-reference-api": "workspace:*",
     "@xwiki/platform-model-remote-url-api": "workspace:*",
