+++ xwiki-platform-core/xwiki-platform-node/src/main/node/core/link-type/link-type-api/api-extractor.json
