+++ xwiki-platform-core/xwiki-platform-node/src/main/node/core/link-type/link-type-api/etc/platform-link-type-api.api.md
@@ -1,4 +1,4 @@
-## API Report File for "@xwiki/platform-link-modal-api"
+## API Report File for "@xwiki/platform-link-type-api"
 
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
