+++ xwiki-platform-core/xwiki-platform-node/src/main/node/core/link-type/link-type-api/package.json
@@ -1,12 +1,12 @@
 {
-  "name": "@xwiki/platform-link-modal-default",
+  "name": "@xwiki/platform-link-type-api",
   "version": "18.8.0-SNAPSHOT",
   "license": "LGPL-2.1-or-later",
   "author": "XWiki Org Community <contact@xwiki.org>",
   "homepage": "https://xwiki.org/",
   "repository": {
     "type": "git",
-    "directory": "xwiki-platform-core/xwiki-platform-node/src/main/node/core/link-modal/link-modal-default",
+    "directory": "xwiki-platform-core/xwiki-platform-node/src/main/node/core/link-type/link-type-api",
     "url": "git+https://github.com/xwiki/xwiki-platform.git"
   },
   "bugs": {
@@ -34,20 +34,13 @@
     "clean": "rimraf dist",
     "lint": "eslint \"./src/**/*.{ts,tsx,vue}\" --max-warnings=0",
     "test": "vitest --run",
-    "typecheck": "vue-tsc"
+    "typecheck": "tsc"
   },
   "types": "./dist/index.d.ts",
   "dependencies": {
-    "@xwiki/platform-dsapi": "workspace:*",
-    "@xwiki/platform-icons": "workspace:*",
-    "@xwiki/platform-link-modal-api": "workspace:*",
-    "@xwiki/platform-link-modal-ui": "workspace:*",
-    "@xwiki/platform-link-suggest-api": "workspace:*",
-    "@xwiki/platform-model-api": "workspace:*",
     "@xwiki/platform-model-reference-api": "workspace:*",
     "@xwiki/platform-model-remote-url-api": "workspace:*",
-    "@xwiki/platform-rendering-api": "workspace:*",
-    "vue-i18n": "catalog:"
+    "@xwiki/platform-rendering-api": "workspace:*"
   },
   "peerDependencies": {
     "inversify": "8.x",
@@ -57,7 +50,6 @@
   "devDependencies": {
     "@microsoft/api-extractor": "catalog:",
     "@xwiki/platform-dev-config": "workspace:*",
-    "@xwiki/platform-fn-utils": "workspace:*",
     "@xwiki/platform-tool-apiextractorconfig": "workspace:*",
     "@xwiki/platform-tool-tsconfig": "workspace:*",
     "@xwiki/platform-tool-tsdocconfig": "workspace:*",
@@ -69,8 +61,7 @@
     "vite": "catalog:",
     "vitest": "catalog:",
     "vitest-mock-extended": "catalog:",
-    "vue": "catalog:",
-    "vue-tsc": "catalog:"
+    "vue": "catalog:"
   },
   "publishConfig": {
     "exports": {
