+++ xwiki-platform-core/xwiki-platform-node/src/main/node/core/link-type/link-type-api/src/data/linkType.ts
@@ -23,7 +23,7 @@
  *
  * The set of valid `type`/`config` combinations is open-ended: it is defined by whichever
  * {@link LinkTargetTypeExtension} implementations are registered (the built-in ones, provided by
- * `@xwiki/platform-link-modal-default`, contribute `"page"`, `"attachment"`, `"url"` and `"email"`), rather than
+ * `@xwiki/platform-link-type-default`, contribute `"page"`, `"attachment"`, `"url"` and `"email"`), rather than
  * being a fixed, closed set of types.
  *
  * @since 18.7.0RC1
