+++ xwiki-platform-core/xwiki-platform-node/src/main/node/core/link-type/link-type-default/etc/platform-link-type-default.api.md
@@ -1,4 +1,4 @@
-## API Report File for "@xwiki/platform-link-modal-default"
+## API Report File for "@xwiki/platform-link-type-default"
 
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
@@ -8,9 +8,9 @@ import { AttachmentReference } from '@xwiki/platform-model-api';
 import { Component } from 'vue';
 import { Container } from 'inversify';
 import { DocumentReference } from '@xwiki/platform-model-api';
-import { LinkTargetReferenceContext } from '@xwiki/platform-link-modal-api';
-import { LinkTargetTypeExtension } from '@xwiki/platform-link-modal-api';
-import { LinkTargetUrlContext } from '@xwiki/platform-link-modal-api';
+import { LinkTargetReferenceContext } from '@xwiki/platform-link-type-api';
+import { LinkTargetTypeExtension } from '@xwiki/platform-link-type-api';
+import { LinkTargetUrlContext } from '@xwiki/platform-link-type-api';
 import { ResourceReference } from '@xwiki/platform-rendering-api';
 
 // @beta
