+++ xwiki-platform-core/xwiki-platform-node/src/main/node/core/link-type/link-type-default/langs/translation-en.json
@@ -1,0 +1,18 @@
+{
+  "link-type.target-types.url.label": "URL",
+  "link-type.target-types.url.url": "URL",
+
+  "link-type.target-types.email.label": "E-mail",
+  "link-type.target-types.email.address": "E-mail address",
+  "link-type.target-types.email.subject": "Message subject",
+  "link-type.target-types.email.body": "Message body",
+
+  "link-type.target-types.page.label": "Page",
+  "link-type.target-types.page.reference": "Page reference",
+  "link-type.target-types.page.query-string": "Query string",
+  "link-type.target-types.page.anchor": "Anchor",
+
+  "link-type.target-types.attachment.label": "Attachment",
+  "link-type.target-types.attachment.reference": "Attachment reference",
+  "link-type.target-types.attachment.query-string": "Query string"
+}
