+++ xwiki-platform-core/xwiki-platform-node/src/main/node/core/link-type/link-type-default/src/extensions/EmailLinkTargetType.ts
@@ -24,7 +24,7 @@ import { tryFallible } from "@xwiki/platform-fn-utils";
 import { ResourceType } from "@xwiki/platform-rendering-api";
 import { injectable } from "inversify";
 import type { LinkEmailConfig } from "../data/linkType";
-import type { LinkTargetTypeExtension } from "@xwiki/platform-link-modal-api";
+import type { LinkTargetTypeExtension } from "@xwiki/platform-link-type-api";
 import type { ResourceReference } from "@xwiki/platform-rendering-api";
 import type { Component } from "vue";
 
@@ -39,7 +39,7 @@ class EmailLinkTargetType implements LinkTargetTypeExtension<LinkEmailConfig> {
   readonly type = "email";
   readonly order = 200;
 
-  getLabel = labelFromTranslations("link-modal.target-types.email.label");
+  getLabel = labelFromTranslations("link-type.target-types.email.label");
 
   createDefaultConfig(): LinkEmailConfig {
     return { address: "" };
