+++ xwiki-platform-core/xwiki-platform-node/src/main/node/core/link-type/link-type-default/src/extensions/__tests__/AttachmentLinkTargetType.test.ts
