+++ xwiki-platform-core/xwiki-platform-node/src/main/node/editors/blocknote-headless/package.json
@@ -48,6 +48,7 @@
     "@xwiki/platform-icons": "workspace:*",
     "@xwiki/platform-link-modal-ui": "workspace:*",
     "@xwiki/platform-link-suggest-api": "workspace:*",
+    "@xwiki/platform-link-type-api": "workspace:*",
     "@xwiki/platform-macros-api": "workspace:*",
     "@xwiki/platform-model-api": "workspace:*",
     "@xwiki/platform-model-reference-api": "workspace:*",
